+++ src/main/java/ladder/generator/BooleanGenerator.java
@@ -1,0 +1,6 @@
+package ladder.generator;
+
+@FunctionalInterface
+public interface BooleanGenerator {
+    boolean generate();
+}
