+++ src/main/java/ladder/generator/RandomBooleanGenerator.java
@@ -1,0 +1,12 @@
+package ladder.generator;
+
+import java.util.Random;
+
+public class RandomBooleanGenerator implements BooleanGenerator {
+    private static final Random RANDOM = new Random();
+
+    @Override
+    public boolean generate() {
+        return RANDOM.nextBoolean();
+    }
+}
