+++ src/main/java/ladder/model/Ladder.java
@@ -1,5 +1,7 @@
 package ladder.model;
 
+import ladder.generator.BooleanGenerator;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
